Replacing CarriageSubsystem.java with Carriage.java
